Rename: Music tab plugin to match new naming conventions
Remove:
Unnecessary music-related json file.
